requirements/requirements-testing.txt
-r requirements-base.txt

# PyTest for running the tests.
pytest==3.0.0
pytest-cov==2.3.1
pytest-django==3.0.0
